remove old testclass and integrated it into the correct placee. Remove sysout from unittest
