+++ src/test/java/spoon/test/ctType/CtTypeParameterTest.java
@@ -1,18 +1,16 @@
 /**
- * Copyright (C) 2006-2018 INRIA and contributors
- * Spoon - http://spoon.gforge.inria.fr/
+ * Copyright (C) 2006-2018 INRIA and contributors Spoon - http://spoon.gforge.inria.fr/
  *
- * This software is governed by the CeCILL-C License under French law and
- * abiding by the rules of distribution of free software. You can use, modify
- * and/or redistribute the software under the terms of the CeCILL-C license as
- * circulated by CEA, CNRS and INRIA at http://www.cecill.info.
+ * This software is governed by the CeCILL-C License under French law and abiding by the rules of
+ * distribution of free software. You can use, modify and/or redistribute the software under the
+ * terms of the CeCILL-C license as circulated by CEA, CNRS and INRIA at http://www.cecill.info.
  *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the CeCILL-C License for more details.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+ * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * CeCILL-C License for more details.
  *
- * The fact that you are presently reading this means that you have had
- * knowledge of the CeCILL-C license and that you accept its terms.
+ * The fact that you are presently reading this means that you have had knowledge of the CeCILL-C
+ * license and that you accept its terms.
  */
 package spoon.test.ctType;
 
@@ -22,7 +20,8 @@ import java.lang.reflect.Method;
 import java.util.List;
 
 import org.junit.Test;
-
+import spoon.FluentLauncher;
+import spoon.processing.AbstractProcessor;
 import spoon.reflect.declaration.CtClass;
 import spoon.reflect.declaration.CtConstructor;
 import spoon.reflect.declaration.CtExecutable;
@@ -219,5 +218,20 @@ public class CtTypeParameterTest {
 			return false;
 		}
 		return thisType.getQualifiedName().equals(thatAdaptedType.getQualifiedName());
+	}
+
+	@Test
+	public void test() {
+		// testcase for issue 3275. The issue was a nullpointer in this code. Was fixed in #3276
+		new FluentLauncher()
+			.inputResource("src/test/resources/issue3275/BOMCostPrice.java")
+			.noClasspath(true)
+			.processor(new AbstractProcessor<CtTypeReference<?>>() {
+					public void process(CtTypeReference<?> element) {
+						// just a operation to enforce the method 
+					String a = element.getSimpleName();
+				}
+			})
+			.buildModel();
 	}
 }
